OP-127: code review

+++ src/main/kotlin/io/openfuture/chain/domain/transaction/TransferTransactionDto.kt
@@ -5,7 +5,7 @@ import io.openfuture.chain.util.TransactionUtils
 
 class TransferTransactionDto(
     timestamp: Long,
-    amount: Long,
+    amount: Double,
     recipientKey: String,
     recipientAddress: String,
     senderKey: String,
