+++ src/main/kotlin/io/openfuture/chain/domain/transaction/data/BaseTransactionData.kt
@@ -4,7 +4,7 @@ import javax.validation.constraints.NotBlank
 import javax.validation.constraints.NotNull
 
 abstract class BaseTransactionData(
-    @field:NotNull var amount: Long? = null,
+    @field:NotNull var amount: Double? = null,
     @field:NotBlank var recipientKey: String? = null,
     @field:NotBlank var recipientAddress: String? = null,
     @field:NotBlank var senderKey: String? = null,
