+++ src/main/resources/db/migration/V1_9__table_wallets.sql
@@ -1,5 +1,5 @@
 CREATE TABLE wallets (
   id            INTEGER PRIMARY KEY,
   address       VARCHAR NOT NULL,
-  balance       REAL
+  balance       DOUBLE
 );
